fix(test): make compat migration fixture build without test helpers
Move securityBuildCandidate and RepositoryRoot into non-test packages so
go build ./... compiles production fixtures that previously referenced
_test.go-only symbols.
